drop old client, types and refactor cache to ids
