+++ src/amocrm/ids/lead_b2c_pipeline_id.py
@@ -1,7 +1,6 @@
 from typing import Literal
 
 from django.conf import settings
-from django.core.cache import cache
 
 from amocrm.dto.pipelines import AmoCRMPipelines
 from amocrm.exceptions import AmoCRMCacheException
@@ -13,16 +12,10 @@ STATUSES_NAMES_TRANSLATE = {
     "purchased": "Успешно реализовано",
     "closed": "Закрыто и не реализовано",
 }
-STATUSES_TO_CACHE = {
-    "unsorted": "amocrm_b2c_unsorted_status_id",
-    "first_contact": "amocrm_b2c_first_contact_status_id",
-    "purchased": "amocrm_b2c_purchased_status_id",
-    "closed": "amocrm_b2c_closed_status_id",
-}
 B2C_PIPELINE = settings.AMOCRM_B2C_PIPELINE_NAME
 
 
-def get_pipeline_status_id_amocrm(status_name: STATUSES_NAMES) -> int:
+def get_b2c_pipeline_status_id(status_name: STATUSES_NAMES) -> int:
     pipelines = [pipeline for pipeline in AmoCRMPipelines.get() if pipeline["name"] == B2C_PIPELINE]
     if len(pipelines) != 1:
         raise AmoCRMCacheException(f"Cannot retrieve {B2C_PIPELINE} pipeline")
@@ -35,9 +28,9 @@ def get_pipeline_status_id_amocrm(status_name: STATUSES_NAMES) -> int:
     raise AmoCRMCacheException(f"Cannot retrieve {status_name} ({ru_status}) status in {B2C_PIPELINE} pipeline")
 
 
-def get_b2c_pipeline_status_id(status_name: STATUSES_NAMES) -> int:
-    cache_key = STATUSES_TO_CACHE[status_name]
-    return cache.get_or_set(cache_key, lambda: get_pipeline_status_id_amocrm(status_name), timeout=None)  # type: ignore
+def get_b2c_pipeline_id() -> int:
+    pipelines = [pipeline for pipeline in AmoCRMPipelines.get() if pipeline["name"] == B2C_PIPELINE]
+    if len(pipelines) != 1:
+        raise AmoCRMCacheException("Cannot retrieve b2c pipeline id")
 
-
-__all__ = ["get_b2c_pipeline_status_id"]
+    return pipelines[0]["id"]
