+++ src/amocrm/tests/http/conftest.py
@@ -2,7 +2,7 @@ import pytest
 
 from django.core.cache import cache
 
-from amocrm.client import AmoCRMClient
+from amocrm.client import AmoCRMHTTP
 
 
 @pytest.fixture(autouse=True)
@@ -12,10 +12,5 @@ def _cached_tokens():
 
 
 @pytest.fixture
-def amocrm_client() -> AmoCRMClient:
-    return AmoCRMClient()
-
-
-@pytest.fixture
-def get(mocker):
-    return mocker.patch("amocrm.client.http.AmoCRMHTTP.get")
+def amocrm_client() -> AmoCRMHTTP:
+    return AmoCRMHTTP()
