+++ src/amocrm/tests/ids/tests_contact_fields_ids.py
@@ -1,0 +1,35 @@
+import pytest
+
+from amocrm.exceptions import AmoCRMCacheException
+from amocrm.ids.contact_fields_ids import get_contact_field_id
+
+
+@pytest.fixture
+def email_field():
+    return dict(
+        id=333,
+        code="EMAIL",
+    )
+
+
+@pytest.fixture
+def fields(email_field):
+    return [email_field, dict(id=123, code="PHONE")]
+
+
+@pytest.fixture(autouse=True)
+def mock_get_contact_fields(mocker, fields):
+    return mocker.patch("amocrm.dto.catalogs.AmoCRMCatalogs.get_contacts_fields", return_value=fields)
+
+
+def test_return_field_id(email_field):
+    got = get_contact_field_id("EMAIL")
+
+    assert got == email_field["id"]
+
+
+def test_fail_if_not_in_cache_and_not_in_response(mock_get_contact_fields):
+    mock_get_contact_fields.return_value = [dict(id=123, code="PHONE")]
+
+    with pytest.raises(AmoCRMCacheException):
+        get_contact_field_id("EMAIL")
